refactor: 주석 제거 및 테스트 패키지 분리
테스트 파일 패키지 분리 및 주석 제거
