+++ src/test/java/bowling/frame/NormalFrameInfoTest.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.frame;
 
 import bowling.entity.frame.NormalFrameInfo;
 import org.junit.jupiter.api.DisplayName;
