+++ src/test/java/bowling/score/MissTest.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.score;
 
 import bowling.entity.Pin;
 import bowling.entity.score.Miss;
